requirements.txt
Flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
anthropic==0.18.1
Werkzeug==3.0.1
gunicorn==21.2.0
dnspython==2.4.2
